feat(staff): Add edit and delete actions
Introduces functionality to edit and delete staff members. This includes updating the UI to display relevant icons and modals for these actions, as well as handling the underlying API calls. Also, improves handling of staff bank details parsing for backward compatibility.

Adds a path alias for `@` to improve module resolution.